+++ CHANGELOG.md
@@ -16,10 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ];
   ```
 - `Logger.getComponentLogger` - replaces `Logger.getSubsystemLogger`
+- `serialize()` and `deserialize()` - serializer that supports cyclic structures (`a['b']['a'] === a`) for sending objects in logs over network
 ### Changed
 - moved `BrowserLogger.ts` to `src/loggers` folder
 - updated `BrowserLogger` documentation
 - Terminology: _subsystem_ => _component_
+- documentation style
+- enhanced documentation
 ### Deprecated
 - `Logger.getSubsystemLogger` - use `Logger.getComponentLogger` instead
 ### Removed